Cleanup API/index
