+++ src/Components/App/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { getMovies } from '../../Utils/API';
+import { getMovies } from '../../Utils/';
 import TitleContainer from '../TitleContainer';
 import { connect } from 'react-redux';
 import { addMovies } from '../../actions/addMovies';
